+++ src/core/configuration.cc
@@ -167,6 +167,9 @@ void Configuration::loadFromInput(std::string&& jsonStr)
         if (m_serverRSAPublicKeyFile.empty()) {
             errorStream << "  Both RSA private and public keys must be provided if provided at all" << std::endl;
         } else {
+            Utils::resolveResidueHomeEnvVar(m_serverRSAPrivateKeyFile);
+            Utils::resolveResidueHomeEnvVar(m_serverRSAPublicKeyFile);
+
             if (!Utils::fileExists(m_serverRSAPrivateKeyFile.c_str()) || !Utils::fileExists(m_serverRSAPublicKeyFile.c_str())) {
                 errorStream << "  RSA private key or public key does not exist" << std::endl;
             } else {
@@ -382,6 +385,7 @@ void Configuration::loadKnownLoggers(const JsonDoc::Value& json, std::stringstre
         }
         std::string easyloggingConfigFile = j.get<std::string>("configuration_file", "");
         if (!easyloggingConfigFile.empty()) {
+            Utils::resolveResidueHomeEnvVar(easyloggingConfigFile);
             if (!Utils::fileExists(easyloggingConfigFile.c_str())) {
                 errorStream << "  File [" << easyloggingConfigFile << "] does not exist" << std::endl;
                 continue;
@@ -532,6 +536,7 @@ void Configuration::loadKnownClients(const JsonDoc::Value& json, std::stringstre
             errorStream << "  RSA public key not provided in known_clients for [" << clientId << "]" << std::endl;
             continue;
         }
+        Utils::resolveResidueHomeEnvVar(publicKey);
         if (m_knownClientsKeys.find(clientId) != m_knownClientsKeys.end()) {
             errorStream << "  Duplicate client ID in known_clients [" << clientId << "]" << std::endl;
             continue;
@@ -604,9 +609,6 @@ void Configuration::loadKnownClients(const JsonDoc::Value& json, std::stringstre
                 for (std::string loggerId : loggerIds) {
                     if (m_knownLoggerUserMap.find(loggerId) == m_knownLoggerUserMap.end()) {
                         m_knownLoggerUserMap.insert(std::make_pair(loggerId, loggerUser));
-
-                        // folowing is only for saving purposes
-                        m_knownClientUserMap.insert(std::make_pair(clientId, loggerUser));
                     } else {
                         // for same user ignore, for different user this is a config warning
                         std::string existingAssignedUser = m_knownLoggerUserMap.at(loggerId);